requirements.txt
diffusers
transformers
accelerate
scipy
torch
torchvision
pillow
fastapi
uvicorn
python-multipart
facenet-pytorch
